requirements.txt
protobuf==3.10.0
pycryptodome==3.20.0
